In tasks 3 & 5 variables was not marked as float

+++ task_1_3.py
@@ -1,3 +1,3 @@
-a = float(input("input cube length:"))
+a = int(input("input cube length:"))
 print("Cube volume:", a*a*a)
 print("Cube face area:", a*a)
